refactor: planetscale.ts db 폴더로 이동
