+++ lib/api.ts
@@ -1,4 +1,4 @@
-import { ViewsTable } from 'lib/planetscale'
+import { ViewsTable } from 'lib/db/planetscale'
 
 /* 💡 Get All View counts */
 export const getAllView = async (): Promise<ViewsTable[]> => {
