+++ pages/api/views/index.ts
@@ -1,4 +1,4 @@
-import { queryBuilder } from 'lib/planetscale'
+import { queryBuilder } from 'lib/db/planetscale'
 import type { NextApiRequest, NextApiResponse } from 'next'
 
 export default async function handler(req: NextApiRequest, res: NextApiResponse) {
